--- server/.env
@@ -1,5 +1,0 @@
-DB_HOST=localhost
-DB_USER=root
-DB_DATABASE=deal
-DB_PASSWORD
-DB_PORT=3306
